Added ViewModel and added the bindings needed for the settings.
Formatted value inside the settings for the font size.

Bound font size for the clock and the setting value to the same variable.

Added converter for int to double value for the size picker.

Implemented clock and bound the format of the clock to the toggle button in the settings.

Removed some empty methods that ended up not being used.